Moved vm::run_no_gdb to the os-independent code
