+++ src/linux/mod.rs
@@ -8,9 +8,7 @@ pub type DebugExitInfo = kvm_bindings::kvm_debug_exit_arch;
 use std::{
 	io,
 	net::{TcpListener, TcpStream},
-	os::unix::prelude::JoinHandleExt,
-	sync::{Arc, Barrier, LazyLock},
-	thread,
+	sync::LazyLock,
 };
 
 use core_affinity::CoreId;
@@ -28,8 +26,6 @@ use crate::{
 		x86_64::kvm_cpu::KvmVm,
 	},
 	serial::Destination,
-	stats::{CpuStats, VmStats},
-	vcpu::VirtualCPU,
 	vm::{UhyveVm, VmResult},
 };
 
@@ -38,7 +34,7 @@ static KVM: LazyLock<Kvm> = LazyLock::new(|| Kvm::new().unwrap());
 /// The signal for kicking vCPUs out of KVM_RUN.
 ///
 /// It is used to stop a vCPU from another thread.
-struct KickSignal;
+pub(crate) struct KickSignal;
 
 impl KickSignal {
 	const RTSIG_OFFSET: libc::c_int = 0;
@@ -51,7 +47,7 @@ impl KickSignal {
 		unsafe { std::mem::transmute(kick_signal) }
 	}
 
-	fn register_handler() -> nix::Result<()> {
+	pub(crate) fn register_handler() -> nix::Result<()> {
 		extern "C" fn handle_signal(_signal: libc::c_int) {}
 		// SAFETY: We don't use the `signal`'s return value.
 		unsafe {
@@ -63,7 +59,7 @@ impl KickSignal {
 	/// Sends the kick signal to a thread.
 	///
 	/// [`KickSignal::register_handler`] should be called prior to this to avoid crashing the program with the default handler.
-	fn pthread_kill(pthread: Pthread) -> nix::Result<()> {
+	pub(crate) fn pthread_kill(pthread: Pthread) -> nix::Result<()> {
 		pthread_kill(pthread, Self::get())
 	}
 }
@@ -79,92 +75,6 @@ impl UhyveVm<KvmVm> {
 			self.run_no_gdb(cpu_affinity)
 		} else {
 			self.run_gdb(cpu_affinity)
-		}
-	}
-
-	fn run_no_gdb(self, cpu_affinity: Option<Vec<CoreId>>) -> VmResult {
-		// After spinning up all vCPU threads, the main thread waits for any vCPU to end execution.
-		let barrier = Arc::new(Barrier::new(2));
-
-		debug!("Starting vCPUs");
-		let threads = self
-			.vcpus
-			.into_iter()
-			.enumerate()
-			.map(|(cpu_id, mut cpu)| {
-				let barrier = barrier.clone();
-				let local_cpu_affinity = cpu_affinity
-					.as_ref()
-					.and_then(|core_ids| core_ids.get(cpu_id).copied());
-
-				thread::spawn(move || {
-					debug!("Create thread for CPU {}", cpu_id);
-					match local_cpu_affinity {
-						Some(core_id) => {
-							debug!("Trying to pin thread {} to CPU {}", cpu_id, core_id.id);
-							core_affinity::set_for_current(core_id); // This does not return an error if it fails :(
-						}
-						None => debug!("No affinity specified, not binding thread"),
-					}
-
-					thread::sleep(std::time::Duration::from_millis(cpu_id as u64 * 50));
-
-					// jump into the VM and execute code of the guest
-					match cpu.run() {
-						Ok((code, stats)) => {
-							if code.is_some() {
-								// Let the main thread continue with kicking the other vCPUs
-								barrier.wait();
-							}
-							(Ok(code), stats)
-						}
-						Err(err) => {
-							error!("CPU {} crashed with {:?}", cpu_id, err);
-							barrier.wait();
-							(Err(err), None)
-						}
-					}
-				})
-			})
-			.collect::<Vec<_>>();
-		debug!("Waiting for first CPU to finish");
-
-		// Wait for one vCPU to return with an exit code.
-		barrier.wait();
-
-		for thread in &threads {
-			KickSignal::pthread_kill(thread.as_pthread_t()).unwrap();
-		}
-
-		let cpu_results = threads
-			.into_iter()
-			.map(|thread| thread.join().unwrap())
-			.collect::<Vec<_>>();
-		let code = match cpu_results
-			.iter()
-			.filter_map(|(ret, _stats)| ret.as_ref().ok())
-			.filter_map(|ret| *ret)
-			.count()
-		{
-			0 => panic!("No return code from any CPU? Maybe all have been kicked?"),
-			1 => cpu_results[0].0.as_ref().unwrap().unwrap(),
-			_ => panic!("more than one thread finished with an exit code (codes: {cpu_results:?})"),
-		};
-
-		let stats: Vec<CpuStats> = cpu_results
-			.iter()
-			.filter_map(|(_ret, stats)| stats.clone())
-			.collect();
-		let output = if let Destination::Buffer(b) = &self.peripherals.serial.destination {
-			Some(String::from_utf8_lossy(&b.lock().unwrap()).into_owned())
-		} else {
-			None
-		};
-
-		VmResult {
-			code,
-			output,
-			stats: Some(VmStats::new(&stats)),
 		}
 	}
